Link on JIRA ticket (optionally)
Summary:
We recently enabled the JIRA integration per {T5422} and found the JIRA comment for every Differential event to be noisy. Since we already get those notifications from Phabricator we didn't want to also get them from JIRA.

But we do want to have a link from the JIRA ticket to the revision and to easily see what its status is. This revision adds an option to use the JIRA API's link feature to do that. It also makes the commenting an option flag.

Test Plan: Tested in our local installation with JIRA On-demand.

Reviewers: btrahan, epriestley, #blessed_reviewers, turadg

Subscribers: jra3, eMxyzptlk, frenchs, aik099, svemir, rmuslimov, cpa199, waynea, epriestley, Korvin, hach-que

Maniphest Tasks: T5422

Differential Revision: https://secure.phabricator.com/D9858